third_party/WebKit/LayoutTests/fast/html/imports/import-expando-gc.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<link rel="import" href="resources/hello.html">
</head>
<body>
<script>
description('Properties in imported dom nodes should be retained.');

var importedDocument;

function getNodeInImport() {
    if (!importedDocument)
        importedDocument = document.querySelector('link').import;
    return importedDocument.querySelector('h1');
}

function test(gcFunction) {
    (function() {
        getNodeInImport().foo = 'Hello';
    })();

    gcFunction();
    shouldBeEqualToString('getNodeInImport().foo', 'Hello');
}

// The minor one should be first. The wrapped node is moved to older space otherwise.
test(GCController.minorCollect.bind(GCController));
test(GCController.collectAll.bind(GCController));

</script>
</body>
</html>
