third_party/WebKit/LayoutTests/fast/html/imports/import-expando-gc-expected.txt

Properties in imported dom nodes should be retained.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getNodeInImport().foo is "Hello"
PASS getNodeInImport().foo is "Hello"
PASS successfullyParsed is true

TEST COMPLETE
